feat: listing saving now working

+++ frontend/src/utils/datefunctions.ts
@@ -4,5 +4,5 @@ export function getYMDString(date: Date | null): string {
 
 export function formatListingDate(dateString: string): string {
   const date = new Date(dateString);
-  return date.toLocaleDateString(undefined, { year: "numeric", month: "long", day: "numeric" });
+  return date.toLocaleDateString(undefined, { year: "numeric", month: "short", day: "numeric" });
 }
